pyproject.toml
[project]
name = "family"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.45.1",
    "streamlit-flow-component>=1.6.1",
    "streamlit-js-eval>=0.1.7",
]
